Add initial command to structuring the project
